morepath_sqlalchemy/requirements.txt
morepath==0.11.1
more.jinja2==0.2
